Rename internal instrumentation module
