+++ instrumentation/internal/internal-application-logger/bootstrap/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+  id("otel.javaagent-bootstrap")
+}
+
+dependencies {
+  compileOnly(project(":javaagent-bootstrap"))
+}
