+++ instrumentation/internal/internal-application-logger/javaagent/build.gradle.kts
@@ -12,6 +12,8 @@ muzzle {
 }
 
 dependencies {
+  bootstrap(project(":instrumentation:internal:internal-application-logger:bootstrap"))
+
   compileOnly(project(":javaagent-bootstrap"))
 
   compileOnly("org.slf4j:slf4j-api") {
