+++ instrumentation/internal/internal-application-logger/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/internal/logging/ApplicationLoggingInstrumentationModule.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.internal.slf4j;
+package io.opentelemetry.javaagent.instrumentation.internal.logging;
 
 import static java.util.Collections.singletonList;
 
@@ -14,10 +14,10 @@ import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import java.util.List;
 
 @AutoService(InstrumentationModule.class)
-public class Slf4jInstrumentationModule extends InstrumentationModule {
+public class ApplicationLoggingInstrumentationModule extends InstrumentationModule {
 
-  public Slf4jInstrumentationModule() {
-    super("internal-slf4j");
+  public ApplicationLoggingInstrumentationModule() {
+    super("internal-application-logging");
   }
 
   @Override
