+++ javaagent-internal-logging-application/build.gradle.kts
@@ -9,10 +9,12 @@ dependencies {
   compileOnly(project(":instrumentation-api"))
   compileOnly(project(":javaagent-bootstrap"))
   compileOnly(project(":javaagent-tooling"))
+  compileOnly(project(":instrumentation:internal:internal-application-logger:bootstrap"))
 
   testImplementation(project(":instrumentation-api"))
   testImplementation(project(":javaagent-bootstrap"))
   testImplementation(project(":javaagent-tooling"))
+  testImplementation(project(":instrumentation:internal:internal-application-logger:bootstrap"))
 
   annotationProcessor("com.google.auto.service:auto-service")
   compileOnly("com.google.auto.service:auto-service-annotations")
