+++ settings.gradle.kts
@@ -185,6 +185,8 @@ hideFromDependabot(":instrumentation:cdi-testing")
 hideFromDependabot(":instrumentation:graphql-java-12.0:javaagent")
 hideFromDependabot(":instrumentation:graphql-java-12.0:library")
 hideFromDependabot(":instrumentation:graphql-java-12.0:testing")
+hideFromDependabot(":instrumentation:internal:internal-application-logger:bootstrap")
+hideFromDependabot(":instrumentation:internal:internal-application-logger:javaagent")
 hideFromDependabot(":instrumentation:internal:internal-class-loader:javaagent")
 hideFromDependabot(":instrumentation:internal:internal-class-loader:javaagent-integration-tests")
 hideFromDependabot(":instrumentation:internal:internal-eclipse-osgi-3.6:javaagent")
@@ -192,7 +194,6 @@ hideFromDependabot(":instrumentation:internal:internal-lambda:javaagent")
 hideFromDependabot(":instrumentation:internal:internal-lambda-java9:javaagent")
 hideFromDependabot(":instrumentation:internal:internal-reflection:javaagent")
 hideFromDependabot(":instrumentation:internal:internal-reflection:javaagent-integration-tests")
-hideFromDependabot(":instrumentation:internal:internal-slf4j:javaagent")
 hideFromDependabot(":instrumentation:internal:internal-url-class-loader:javaagent")
 hideFromDependabot(":instrumentation:internal:internal-url-class-loader:javaagent-integration-tests")
 hideFromDependabot(":instrumentation:c3p0-0.9:javaagent")
